requirements.txt
Flask
Flask-CORS
pymongo
python-dotenv
cryptography
gunicorn